Extract PackageContainer into separate file
